requirements.txt
python-dotenv==0.19.2
pyarrow==7.0.0
lightgbm==3.2.1
numerapi==2.9.4
pandas==1.4.1
scikit-learn==0.24.2
scipy==1.8.0
matplotlib==3.5.1
ipykernel==6.9.1
tabulate==0.8.9